+++ pypom_navigation/navigation.py
@@ -55,7 +55,7 @@ class Navigation(object):
         """
         kwargs = self.merge_kwargs(kwargs)
         page_url = urljoin(self.skin_base_url, self.get_page_url(page_id))
-        page_instance = self._page_instance(page_id=page_id, **kwargs)
+        page_instance = self.get_page_instance(page_id=page_id, **kwargs)
         page_instance.driver.visit(page_url)
         page_instance.wait_for_page_to_load()
         self.setPage(page_instance, page_id=page_id)
@@ -66,7 +66,7 @@ class Navigation(object):
             mapped to the passed page_id
         """
         kwargs = self.merge_kwargs(kwargs)
-        page_instance = self._page_instance(page_id=page_id, **kwargs)
+        page_instance = self.get_page_instance(page_id=page_id, **kwargs)
         page_instance.wait_for_page_to_load()
         self.setPage(page_instance, page_id=page_id)
         return page_instance
@@ -81,7 +81,7 @@ class Navigation(object):
         if page_id:
             return self.update_page(page_id, **kwargs)
         else:
-            page_instance = self._page_instance(fallback=fallback, **kwargs)
+            page_instance = self.get_page_instance(fallback=fallback, **kwargs)
             page_instance.wait_for_page_to_load()
             self.setPage(page_instance)
             return page_instance
@@ -107,7 +107,8 @@ class Navigation(object):
             page_id=page_id,
             fallback=fallback)
 
-    def _page_instance(self, page_id=None, fallback=None, **kwargs):
+    def get_page_instance(self, page_id=None, fallback=None, **kwargs):
+        """ Get a fresh page instance """
         page_class = self.get_page_class(page_id=page_id, fallback=fallback)
         return page_class(self.driver, **kwargs)